Update retry logic for resiliency
This is based on a real-life scenario where the Unreal Engine HTTP layer
got into a weird state where it would send the payload for HTTP requests
successfully, but failed to parse/process/notify on the HTTP response.
So the cloud service was receiving the payload successfully every time
and sending HTTP 200, but the plugin thought it was timing out. Increase
the timeout threshold and make sure that retried payloads are the same
so that they can be successfully dedup'd in this situation.

* When retrying after a previous failure, make sure to only attempt to
  process the same size payload as before; this ensures any ingest dedup
  logic can dedup the request if needed.
* Added RetrySamePayloadSize automation test to cover new retry logic.
* Retry interval will increase with each consecutive failure up to a max
  of 5 minutes
* Default request timeout increased to 90 seconds
* Default retry interval increased to 30 seconds
* Log the expected retry interval for clarity