Add auto-flush after N bytes logged feature
Trigger immediate flush of logs if at least N bytes has been logged and
its been at least a certain amount of time since the last flush.

This ensures that we still don't trigger a log flush too often (limit
CPU overhead of logging plugin) while also triggering an immediate
flush if we have a material amount of pending data.

This is helpful for the Client launch mode where we are not sending
log data but we are sending metrics/analytics data, and if we queue
up a certain amount of analytics data we want to send it sooner
rather than waiting for the long default processing interval (10 min
for client launch configuration)

Also increase max line length to 512KB now that multi-line messages
are encoded in the backing file as a single line of text.

Also fix a rare race condition crash where the post-processing HTTP
callback that referenced the Streamer after the Streamer was cleaned
up caused a SIGSEGV.

Fixed a bug with the raw JSON encoder that would cause the raw
JSON to appear after the message event tag.

Fixed an issue where Streamer would abort with a fatal error if
the backing log file didn't exist when it started. Treat this situation
as if the backing file existed but had zero bytes.

Added automation tests to cover all new scenarios.