Fix unit tests failing due to cached progress markers
When running the test suite many times within an editor session
tests would start to randomly fail with weird reasons.
The root cause was that even though the state INI
file was being deleted, UE was remembering the previous
state in the config cache. Clearing the config cache
for the test state file now fixes this problem!